+++ AppCore/ApplicationState.py
@@ -1,4 +1,5 @@
 class ApplicationState:
+    
     @property
     def can_publish_staged_resources(self) -> bool:
         return NotImplemented